Change thirdOfArmor to int
Compiler was throwing an error since the modulo operator
only works with integers, not doubles.

+++ CrazyRandomSword.cpp
@@ -10,7 +10,7 @@
 double CrazyRandomSword::hit(double armor) {
     srand(time(0));                 // seed random number generator
 
-    double thirdOfArmor = armor / 3;
+    int thirdOfArmor = armor / 3;
 
     // ignoredArmor is random number between 2 - thirdArmor
     int ignoredArmor = static_cast<int>(rand() % (thirdOfArmor - 1) + 2);
